add a "source code" tokenizer.
This commit introduces a few interfaces and classes for the tokenizer to
operate.

The lexer is a regex-based thingy that is matched from a certain
position within the "source code" string.

Everything in the lexer is a static reference, meaning that you don't
need to create new instances of anything in order to use them.