update Cell tokenizer, SYNTAX file and others.
This commit cleans up some stuff in the tokenizer, the Cell definition
doesn't have the static members defined anymore and the lexer uses only
one instance of the class.

The parser is also updated to match the new format of cells and ranges.

Previously, a cell was composed of letters and numbers, however, due to
design issues I thought it was better to use numbers only and represent
them like:

    $0 (for cells) and $0:1 (for ranges)

This commit also cleans up both the CellContext and ICellContext types
because of the removal of the letter part in the Cell definition, types
like CellPosition and RangePositions were removed as well.

Next commits will focus on the implementation of ICellContext so users
can actually execute some code from the repl at least.