add basic function support.
This commit adds basic operations regarding to functions.

The programming model of Cell uses a single string for error messages,
so most stuff was updated to use that error string.

The only time you should see an exception popping out is when something
is wrong on the C# end, for example: Trying to insert duplicate names in
the global function table would throw an exception.

Null or invalid identifiers are also part of the gang, if a null value
is spotted, ArgumentNullException is thrown.

Anyway, this also updates the IExpression.Evaluate method to comply with
the pattern.

That's it, next commits should allow some functions to work with Cell.